# api/v1alpha1/types.gen.go
// Package v1alpha1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package v1alpha1

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ApplicationService.
const (
	Container ApplicationService = "container"
	Webserver ApplicationService = "webserver"
)

// Application defines model for Application.
type Application struct {
	// Name Name of the application
	Name string `json:"name"`

	// Service Service of the application
	Service ApplicationService `json:"service"`

	// Tier Policy Tier of the application
	Tier *int `json:"tier,omitempty"`

	// Zones Zones of the application
	Zones *[]string `json:"zones,omitempty"`
}

// ApplicationService Service of the application
type ApplicationService string

// ApplicationList defines model for ApplicationList.
type ApplicationList struct {
	Applications []ApplicationResponse `json:"applications"`

	// NextPageToken Token for retrieving the next page of results
	NextPageToken *string `json:"next_page_token,omitempty"`
}

// ApplicationResponse defines model for ApplicationResponse.
type ApplicationResponse struct {
	// Id ID of the application
	Id *openapi_types.UUID `json:"id,omitempty"`

	// Name Name of the application
	Name *string `json:"name,omitempty"`

	// Service Service of the application
	Service *string `json:"service,omitempty"`

	// Tier Policy Tier of the application
	Tier *int `json:"tier,omitempty"`

	// Zones Zones of the application
	Zones *[]string `json:"zones,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *int `json:"code,omitempty"`

	// Error Error message
	Error string `json:"error"`
}

// CreateApplicationParams defines parameters for CreateApplication.
type CreateApplicationParams struct {
	// Id Optional ID for the application
	Id *string `form:"id,omitempty" json:"id,omitempty"`
}

// CreateApplicationJSONRequestBody defines body for CreateApplication for application/json ContentType.
type CreateApplicationJSONRequestBody = Application